fix CamDepth(WithDepth)

+++ CamDemo(WithDepth).py
@@ -133,7 +133,7 @@ with dai.Device() as device:
                 )
                 cv2.putText(
                     filtered,
-                    "x:{:.1f}m".format(grids[i][j][1].astype(np.uint8)),
+                    "x:{:.1f}m".format(grids[i][j][1]),
                     (
                         int(filtered.shape[1] * j / GRID_NUM_W) + 3,
                         int(filtered.shape[0] * i / GRID_NUM_H) + 24,
@@ -144,7 +144,7 @@ with dai.Device() as device:
                 )
                 cv2.putText(
                     filtered,
-                    "y:{:.1f}m".format(grids[i][j][2].astype(np.uint8)),
+                    "y:{:.1f}m".format(grids[i][j][2]),
                     (
                         int(filtered.shape[1] * j / GRID_NUM_W) + 3,
                         int(filtered.shape[0] * i / GRID_NUM_H) + 36,
@@ -155,7 +155,7 @@ with dai.Device() as device:
                 )
                 cv2.putText(
                     filtered,
-                    "z:{:.1f}m".format(grids[i][j][3].astype(np.uint8)),
+                    "z:{:.1f}m".format(grids[i][j][3]),
                     (
                         int(filtered.shape[1] * j / GRID_NUM_W) + 3,
                         int(filtered.shape[0] * i / GRID_NUM_H) + 48,
